+++ core/components/tabs/tabs-nav-viewport.vue
@@ -7,11 +7,30 @@
 </template>
 
 <script>
+import ResizeObserver from 'resize-observer-polyfill'
+import debounce from 'lodash.debounce'
 export default {
 	name: 'zov-tabs-nav-viewport',
 	props: {
 		scrollX: Number,
 		scrollY: Number
+	},
+	mounted() {
+		// 设置resize观察者
+		this.resizeObserver = new ResizeObserver(
+			debounce(this.onResize.bind(this), 300)
+		)
+		// 观察viewport变化
+		this.resizeObserver.observe(this.$refs.viewport)
+		this.resizeObserver.observe(this.$refs.roller)
+	},
+	beforeDestroy() {
+		this.resizeObserver.disconnect()
+	},
+	data() {
+		return {
+			resizeObserver: null
+		}
 	},
 	computed: {
 		styles() {
@@ -23,6 +42,9 @@ export default {
 		}
 	},
 	methods: {
+		onResize() {
+			this.$emit('resize')
+		},
 		getRollerRect() {
 			return this.$refs.roller.getBoundingClientRect()
 		},